backend/requirements.txt
# Web3 and Ethereum
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1

# FastAPI for API server
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.12.3

# Environment variables
python-dotenv==1.0.0

# IPFS (optional, for CID handling)
ipfshttpclient==0.8.0a2

# Utilities
requests==2.31.0
openai==2.5.0
